feat: enhance contract parsing with task-based LLM integration and OCR cleanup
Major updates:
- Add Docling OCR cleanup service for post-processing scanned documents
- Enhance task-based parser with chunk task tagging and progress tracking
- Refactor system config service with dynamic LLM/embedding configuration
- Add unified contract upload component with Markdown preview
- Add client extended fields migration and regenerate Prisma schema
- Update GraphQL schema with new parsing and RAG-related types
- Add case study generation modal component

+++ apps/api/src/llm-parser/entities/contract-type-detection.entity.ts
@@ -1,4 +1,4 @@
-import { Field, ObjectType, Int } from '@nestjs/graphql';
+import { Field, ObjectType, Int, Float } from '@nestjs/graphql';
 
 /**
  * 合同类型检测结果
@@ -8,7 +8,7 @@ export class ContractTypeDetectionResult {
   @Field(() => String, { nullable: true, description: '检测到的合同类型' })
   detectedType!: string | null;
 
-  @Field(() => Int, { description: '置信度 (0-1)' })
+  @Field(() => Float, { description: '置信度 (0-1)' })
   confidence!: number;
 
   @Field(() => String, { description: '判断依据说明' })
